Factor out SigningTestCase for reuse
Allows sharing a valid certificate setup with SignerTest. In SignerTest,
reuse good certificate as the main certificate for signature and add a
second certificate for testSignWithMultiCertificate.
